+++ testapps/nextjs/.wundergraph/wundergraph.config.ts
@@ -90,7 +90,7 @@ const graphQLAPI = introspect.graphql({
 
 const spaceX = introspect.graphql({
 	apiNamespace: 'spacex',
-	url: 'https://api.spacex.land/graphql/',
+	url: 'https://spacex-api.fly.dev/graphql/',
 });
 
 const counter = introspect.graphql({